Introduce "rename" command.
This commit also includes various improvements and tidy-ups to the way
command options and reporting options are implemented.

Command options now take a callable as an argument to allow arbitrary
code to be called when an option is encountered. This allows e.g. bit
fields to be ORed into, or enum variables to be set to specific values,
when given options are encountered, rather being restricted just to
setting a boolean to true.