Refactor TimeLog to use a simpler load-edit-save cycle.
Instead of different write operations writing to the file in different
ways, we now have a single save operation, and each of the different
write operations modifies the in-memory data structure and then triggers
saving of the entire structure to the file in the same way each time.

Also use Transaction class with RAII to ensure cache is marked as stale
if write operation does not complete.